Remove --gif option & enable gif support for --folder
* gif file are no autodetect
* enable the gif support for --folder #30
