+++ main.py
@@ -8,8 +8,7 @@ import argv
 from config import Config as conf
 from utils import setup_log, read_image, check_shape
 
-from processing.gif import SimpleGIFTransform
-from processing.image import SimpleImageTransform, MultipleImageTransform, FolderImageTransform
+from processing import SimpleTransform, FolderImageTransform, MultipleImageTransform
 from transform.gan.mask import CorrectToMask, MaskrefToMaskdet, MaskfinToNude
 from transform.opencv.resize import ImageToCrop, ImageToOverlay, ImageToRescale, ImageToResized, ImageToResizedCrop
 from transform.opencv.correct import DressToCorrect
@@ -89,11 +88,7 @@ def select_processing():
     """
     phases = select_phases()
     if conf.args['folder']:
-      process = processing_image_folder(phases)
-    elif conf.args['gif'] and conf.args['n_runs'] != 1:
-        process = multiple_gif_processing(phases, conf.args['n_runs'])
-    elif conf.args['gif']:
-        process = simple_gif_processing(phases)
+        process = processing_image_folder(phases)
     elif conf.args['n_runs'] != 1:
         process = multiple_image_processing(phases, conf.args['n_runs'])
     else:
@@ -102,38 +97,13 @@ def select_processing():
     return process
 
 
-def simple_gif_processing(phases):
-    """
-    Define a simple gif process ready to run
-    :param phases: <ImageTransform[]> list of image transformation
-    :return: <SimpleGIFTransform> a gif process run ready
-    """
-    return SimpleGIFTransform(conf.args['input'], phases, conf.args['output'])
-
-
-def multiple_gif_processing(phases, n):
-    """
-    Define a multiple gif process ready to run
-    :param phases: <ImageTransform[]> list of image transformation
-    :param n: number of times to process
-    :return: <MultipleTransform> a multiple gif process run ready
-    """
-    filename, extension = os.path.splitext(conf.args['output'])
-    return MultipleImageTransform(
-        [conf.args['input'] for _ in range(n)],
-        phases,
-        ["{}{}{}".format(filename, i, extension) for i in range(n)],
-        SimpleGIFTransform
-    )
-
-
 def simple_image_processing(phases):
     """
     Define a simple image process ready to run
     :param phases: <ImageTransform[]> list of image transformation
-    :return: <SimpleImageTransform> a image process run ready
+    :return: <SimpleTransform> a image process run ready
     """
-    return SimpleImageTransform(conf.args['input'], phases, conf.args['output'])
+    return SimpleTransform(conf.args['input'], phases, conf.args['output'])
 
 
 def multiple_image_processing(phases, n):
@@ -150,13 +120,15 @@ def multiple_image_processing(phases, n):
         ["{}{}{}".format(filename, i, extension) for i in range(n)]
     )
 
+
 def processing_image_folder(phases):
     """
     Define a folder image process ready to run
     :param phases: <ImageTransform[]> list of image transformation
-    :return: <SimpleImageTransform> a image process run ready
+    :return: <FolderImageTransform> a image process run ready
     """
     return FolderImageTransform(conf.args['input'], phases, conf.args['output'])
+
 
 if __name__ == "__main__":
     freeze_support()
